+++ pkg/redo/writer/memory/mem_log_writer_test.go
@@ -19,7 +19,6 @@ import (
 
 	"github.com/pingcap/ticdc/pkg/common"
 	pevent "github.com/pingcap/ticdc/pkg/common/event"
-	"github.com/pingcap/ticdc/pkg/config"
 	"github.com/pingcap/ticdc/pkg/redo"
 	"github.com/pingcap/ticdc/pkg/redo/writer"
 	"github.com/pingcap/ticdc/pkg/util"
@@ -61,35 +60,6 @@ func TestWriteDML(t *testing.T) {
 	testWriteEvents(t, ddls)
 }
 
-func TestNewLogWriterUsesNormalizedStorageConfig(t *testing.T) {
-	t.Parallel()
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	_, uri, err := util.GetTestExtStorage(ctx, t.TempDir())
-	require.NoError(t, err)
-
-	cfg, err := writer.NewConfig(
-		common.NewChangeFeedIDWithName("test-changefeed", common.DefaultKeyspaceName),
-		&config.ConsistentConfig{
-			MaxLogSize:        util.AddressOf(int64(10)),
-			FlushIntervalInMs: util.AddressOf(int64(redo.DefaultFlushIntervalInMs)),
-			EncodingWorkerNum: util.AddressOf(redo.DefaultEncodingWorkerNum),
-			FlushWorkerNum:    util.AddressOf(redo.DefaultFlushWorkerNum),
-			Storage:           util.AddressOf(uri.String()),
-			FlushConcurrency:  util.AddressOf(1),
-		},
-	)
-	require.NoError(t, err)
-
-	lw, err := NewLogWriter(ctx, cfg, redo.RedoDDLLogFileType)
-	require.NoError(t, err)
-	require.True(t, cfg.UseExternalStorage())
-	require.Equal(t, uri.String(), cfg.URI().String())
-	require.NoError(t, lw.Close())
-}
-
 func testWriteEvents(t *testing.T, events []writer.RedoEvent) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -98,14 +68,7 @@ func testWriteEvents(t *testing.T, events []writer.RedoEvent) {
 	require.NoError(t, err)
 	lwcfg, err := writer.NewConfig(
 		common.NewChangeFeedIDWithName("test-changefeed", common.DefaultKeyspaceName),
-		&config.ConsistentConfig{
-			MaxLogSize:        util.AddressOf(int64(10)),
-			FlushIntervalInMs: util.AddressOf(int64(redo.DefaultFlushIntervalInMs)),
-			EncodingWorkerNum: util.AddressOf(redo.DefaultEncodingWorkerNum),
-			FlushWorkerNum:    util.AddressOf(redo.DefaultFlushWorkerNum),
-			Storage:           util.AddressOf(uri.String()),
-			FlushConcurrency:  util.AddressOf(1),
-		},
+		newTestConsistentConfig(uri.String()),
 	)
 	require.NoError(t, err)
 	filename := t.Name()
